move the batch_request_contains_batch_data() check to a common location
