+++ great_expectations/checkpoint/configurator.py
@@ -5,11 +5,13 @@ from typing import Dict, List, Optional, Union
 from ruamel.yaml.comments import CommentedMap
 
 from great_expectations.checkpoint.util import (
-    batch_request_contains_batch_data,
     batch_request_in_validations_contains_batch_data,
     get_validations_with_batch_request_as_dict,
 )
-from great_expectations.core.batch import get_batch_request_as_dict
+from great_expectations.core.batch import (
+    batch_request_contains_batch_data,
+    get_batch_request_as_dict,
+)
 from great_expectations.data_context.types.base import (
     CheckpointConfig,
     checkpointConfigSchema,
