+++ great_expectations/rule_based_profiler/toolkit.py
@@ -2,7 +2,11 @@ import os
 from typing import List, Optional, Union
 
 import great_expectations.exceptions as ge_exceptions
-from great_expectations.checkpoint.util import batch_request_contains_batch_data
+from great_expectations.core.batch import (
+    BatchRequest,
+    RuntimeBatchRequest,
+    batch_request_contains_batch_data,
+)
 from great_expectations.data_context.store import ProfilerStore
 from great_expectations.data_context.types.base import DataContextConfigDefaults
 from great_expectations.data_context.types.resource_identifiers import (
@@ -21,7 +25,7 @@ def add_profiler(
     profiler_store: ProfilerStore,
     ge_cloud_id: Optional[str] = None,
 ) -> RuleBasedProfiler:
-    if not _check_validity_of_batch_requests_in_config(config):
+    if not _check_validity_of_batch_requests_in_config(config=config):
         raise ge_exceptions.InvalidConfigError(
             f'batch_data found in batch_request cannot be saved to ProfilerStore "{profiler_store.store_name}"'
         )
@@ -55,22 +59,26 @@ def _check_validity_of_batch_requests_in_config(
     config: RuleBasedProfilerConfig,
 ) -> bool:
     # Evaluate nested types in RuleConfig to parse out BatchRequests
-    batch_requests = []
+    batch_requests: List[Union[BatchRequest, RuntimeBatchRequest, dict]] = []
+    rule: dict
     for rule in config.rules.values():
 
-        domain_builder = rule["domain_builder"]
+        domain_builder: dict = rule["domain_builder"]
         if "batch_request" in domain_builder:
             batch_requests.append(domain_builder["batch_request"])
 
-        parameter_builders = rule.get("parameter_builders", [])
+        parameter_builders: List[dict] = rule.get("parameter_builders", [])
+        parameter_builder: dict
         for parameter_builder in parameter_builders:
             if "batch_request" in parameter_builder:
                 batch_requests.append(parameter_builder["batch_request"])
 
     # DataFrames shouldn't be saved to ProfilerStore
+    batch_request: Union[BatchRequest, RuntimeBatchRequest, dict]
     for batch_request in batch_requests:
         if batch_request_contains_batch_data(batch_request=batch_request):
             return False
+
     return True
